disable possibility to edit apps with custom edito
