+++ apps/chat/src/components/Common/ActionButton.tsx
@@ -1,0 +1,38 @@
+import { TablerIconsProps } from '@tabler/icons-react';
+
+import { useTranslation } from 'next-i18next';
+
+import { Translation } from '@/src/types/translation';
+
+import Tooltip from './Tooltip';
+
+interface Props {
+  name: string;
+  dataQa: string;
+  disabled?: boolean;
+  Icon?: (props: TablerIconsProps) => JSX.Element;
+  onClick?: (e: React.MouseEvent) => void;
+}
+
+export const ActionButton: React.FC<Props> = ({
+  name,
+  dataQa,
+  disabled,
+  Icon,
+  onClick,
+}) => {
+  const { t } = useTranslation(Translation.Common);
+
+  return (
+    <Tooltip tooltip={t(name)}>
+      <button
+        disabled={disabled}
+        onClick={onClick}
+        className="icon-button"
+        data-qa={dataQa}
+      >
+        {Icon && <Icon className="size-6" strokeWidth="1.5" />}
+      </button>
+    </Tooltip>
+  );
+};
