+++ apps/chat/src/pages/apps-editor/[slug]/settings/index.tsx
@@ -96,7 +96,11 @@ export default function AppsSettings() {
                   ] ?? '')
                 : decode(slug.toString())
             }
-            hasCustomEditor={!!schema?.['dial:applicationTypeEditorUrl']}
+            hasCustomEditor={
+              !!schema?.[
+                ApplicationTypeSchemaProperties.applicationTypeEditorUrl
+              ]
+            }
           />
           <div className="flex size-full grow overflow-hidden">
             {applicationData && (
